Revert "BUG/MEDIUM: resolvers: always check a valid item in query_list"
This reverts commit f701ae71200de2f9af1670a4f10dbee54d402f99.

Finally it was found to cause a regression in deinit() when a server uses
a CNAME. A corruption in the list makes dns_purge_resolution_query_items()
spin over itself at the end of the list. A simple reproducer is:

  resolvers dns
    nameserver cf1 1.0.0.1:53

  listen test
    bind 127.0.0.1:443
    server s master.haproxy.org:443 resolvers dns

Sending it SIGUSR1 makes it loop. It definitely indicates that there
are other strange things in this area but this code is so fragile that
we'd better leave the previous problem than its fix that makes the
situation worse for those reloading often. 2.2 doesn't have this
problem with the same fix.