BUG/MEDIUM: cli: Never wait for more data on client shutdown
When a shutdown is detected on the cli, we try to execute all pending
commands first before closing the connection. It is required because
commands execution is serialized. However, when the last part is a partial
command, the cli connection is not closed, waiting for more data. Because
there is no timeout for now on the cli socket, the connection remains
infinitely in this state. And because the maxconn is set to 10, if it
happens several times, the cli socket quickly becomes unresponsive because
all its slots are waiting for more data on a closed connections.

This patch should fix the issue #1512. It must be backported as far as 2.0.

(cherry picked from commit 0f727dabf51d4ffead40fd43feb7c07193ebde99)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 34cf2ca548d238b53e079cfc48fabca1b2aad3ca)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit c8c4e3f2ad8fed5c5dba3ce3f3b2bbbe0234332e)
Signed-off-by: Willy Tarreau <w@1wt.eu>