BUG/MINOR: http: fix recent regression on authorization in legacy mode
The backport to 2.0 of commit 1df3c305b ("BUG/MINOR: http: Authorization
value can have multiple spaces after the scheme") was not correct for the
legacy part, it used to maintain the "p+1" start offset to skip the first
character expected to be a space, except that the spaces are all trimmed
by the previous loop. Thus the base64 part misses its first character and
cannot be decoded anymore.

Special thanks to Veiko Kukk and Lukas Tribus for reporting and bisecting
this bug with a reproducer. This was tracked as issue #1516.

There is no upstream commit since it's a fix for an issue affecting 2.0
only.