BUG/MEDIUM: mcli: do not try to parse empty buffers
When pcli_parse_request() is called with an empty buffer, it still tries
to parse it and can go on believing it finds an empty request if the last
char before the beginning of the buffer is a '\n'. In this case it overwrites
it with a zero and processes it as an empty command, doing nothing but not
making the buffer progress. This results in an infinite loop that is stopped
by the watchdog. For a reason related to another issue (yet to be fixed),
this can easily be reproduced by pipelining lots of commands such as
"show version".

Let's add a length check after the search for a '\n'.

This needs to be backported as far as 2.0.

(cherry picked from commit 6cd93f52e974aca4dac7aa2756c4e17659ae5354)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit c67d764d8ae636c378951243fa4038e2dedbf7a8)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 9b736ebace8402e7969aad52d33a1282058a8402)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 70f3dc034562e5330adf0890d647ce376948b99a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 439bc396dfda1bc839f52f306ebe0c623ceba1f9)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>